pyproject.toml
[tool.poetry]
name = "testtasl"
version = "0.1.0"
description = ""
authors = ["hoolywant <sir.saltickow@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
